Fix: In the example topology I made a loop with the switches and did not enabled the spanning tree algorithm.
Fix: TopologyDB was trying to get hosts.gateway attribute but it does not exist when you dont use my special minirouter to create the topology.

Fix: in minigen changed host.cmd for host.cmdPrint to get what is going on in in the flowserver. Maybe I should log everything into a file so we dont flood the screen???
Signed-off-by: edgarcosta92 <cedgar@student.ethz.ch>